Brush up the Heatmaps

+++ service/jfr-heatmap-generator/src/main/java/pbouda/jeffrey/generator/heatmap/HeatmapModel.java
@@ -2,5 +2,5 @@ package pbouda.jeffrey.generator.heatmap;
 
 import com.fasterxml.jackson.databind.node.ArrayNode;
 
-public record HeatmapModel(int maxvalue, ArrayNode series) {
+public record HeatmapModel(long maxvalue, ArrayNode series) {
 }
